Migrate to Mojo v1.0.0rc0
- mojo pin ==1.0.0b2 -> ==1.0.0rc0; json/mozz/mojodoc deps stay pinned to
  branch=main (json git tag v0.2.1 and mozz git tag v0.1.4 dropped -- no
  post-rc0 release cut yet for either)
- std.memory.Span -> std.collections.span.Span (71 files); Span(...)
  constructor keyword ptr= -> unsafe_ptr= (19 call sites)
- OwnedDLHandle.get_function no longer returns a field-storable callable --
  new dl_sym() helper in flare/utils/dylib.mojo (get_symbol + raw-address
  Pointer.unsafe_bitcast) replaces 100+ call sites across FFI-heavy modules
  (crypto/hmac, http/fs, net/socket, http/encoding, tls/*, tcp/stream,
  ws/*, quic/crypto, http/middleware); knock-on `raises` added throughout
  the call chains, and __del__/close() cleanup paths wrapped in
  try/except: pass since destructors can't be raises
- List.capacity field -> .capacity() method (17 sites)
- InlineArray comptime lookup table needs materialize[...]() before
  indexing (http/hpack_huffman_simd.mojo)
- self-referencing String reassignment (x = String(x.strip())) now trips
  the aliasing checker -- fixed via an intermediate variable
  (http/multipart.mojo, ws/extensions.mojo)
- perf_counter_ns() returns Int, not UInt (http/metrics.mojo,
  http/reliability.mojo, tests/runtime/test_libc_time.mojo)
- String positional indexing s[i] -> s[byte=i] (http/_reactor/keepalive_scan.mojo,
  http/session.mojo, fuzz/fuzz_extended_connect.mojo)
- os.path.remove -> os.remove (tests/uds/test_uds_listener.mojo)
- field 'children'/'array_val'-style recursive struct needs an explicit
  no-op __deinit__ to satisfy Deinitable (http/template.mojo)
- misc Int/UInt/UInt8 coercion tightening; decompress_gzip error-message
  slice bound

+++ fuzz/fuzz_extractors.mojo
@@ -65,7 +65,12 @@ def _to_ascii(data: List[UInt8], start: Int, end: Int) -> String:
     var out = String(capacity=n)
     for i in range(start, end):
         var b = data[i]
-        if b >= 32 and b < 127 and b != Int(ord("?")) and b != Int(ord("#")):
+        if (
+            b >= 32
+            and b < 127
+            and b != UInt8(ord("?"))
+            and b != UInt8(ord("#"))
+        ):
             out += chr(Int(b))
         else:
             out += "_"
