+++ tests/http/test_reliability.mojo
@@ -277,7 +277,7 @@ def test_retry_backoff_sleeps_between_attempts() raises:
     # No assertion on a strict lower bound -- jitter can be 0 --
     # but the upper bound must be sane: 3 attempts * 10 ms cap +
     # generous noise.
-    assert_true(elapsed_ms < UInt(2_000))
+    assert_true(elapsed_ms < 2_000)
     _free_counter(addr)
